[NumPy] Remove references to deprecated NumPy type aliases.
This change replaces references to a number of deprecated NumPy type aliases (np.bool, np.int, np.float, np.complex, np.object, np.str) with their recommended replacement (bool, int, float, complex, object, str).

NumPy 1.24 drops the deprecated aliases, so we must remove uses before updating NumPy.

PiperOrigin-RevId: 496401775

+++ tensorflow_datasets/core/utils/dtype_utils_test.py
@@ -61,9 +61,9 @@ def test_dtype(fn, dtype, result):
 def test_is_np_sub_dtype():
   assert dtype_utils.is_np_sub_dtype(np.int32, np.integer)
   assert dtype_utils.is_np_sub_dtype(np.int64, np.integer)
-  assert dtype_utils.is_np_sub_dtype(np.float, np.floating)
+  assert dtype_utils.is_np_sub_dtype(float, np.floating)
   assert not dtype_utils.is_np_sub_dtype(np.int64, np.floating)
-  assert not dtype_utils.is_np_sub_dtype(np.float, np.integer)
+  assert not dtype_utils.is_np_sub_dtype(float, np.integer)
 
 
 def test_is_same_dtype_type():
